🚚 Rename: compoents 하위에 공통, 페이지별 컴포넌트 폴더 생성 및 이동
