+++ components/boards/BestPost/index.tsx
