+++ components/boards/BestPost/style.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/';
+@import '../../../styles/';
 
 .container {
   flex: 1;
