+++ components/boards/BoardPost/index.tsx
