+++ components/boards/BoardPost/style.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/';
+@import '../../../styles/';
 
 .container {
   border-bottom: 1px solid $cool-gray200;
