+++ components/common/Button/style.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/';
+@import '../../../styles/';
 
 .button {
   cursor: pointer;
