+++ components/common/Container/style.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles';
+@import '../../../styles';
 
 .container {
   max-width: $max-width-size;
