+++ components/common/SearchBar/style.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/';
+@import '../../../styles/';
 
 .wrapper {
   display: flex;
