+++ components/common/TitleText/style.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles';
+@import '../../../styles';
 
 .title {
   @include font-base($cool-gray900, 700, 20px, 24px);
