+++ components/common/TopNavigation/style.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles';
+@import '../../../styles';
 
 .header {
   display: flex;
